Add a script to automatically update Git repos
Added a script to run updates in $HOME/workspace for Git repositories in
order to either manually invoke the script or schedule the updates
periodically with cron.  This will save me some time down the road.  The
shell script will be symlinked to /usr/local/bin/git-update-repos.

The script requires the existence of $HOME/workspace.  More robustness
can be added later as deemed necessary.

The largest limitation is dealing with the Git "master" branch naming
convention change, which was a joke, but needs to be accounted for.
Currently I am also tracking "main" and "trunk" branches.